Modified restaurant schema to include ownerId. Modified review schema to include edit logic and owner response

+++ api/data-gen.js
@@ -7,6 +7,7 @@ const User = require('./schema_models/userSchema.js');
 const Restaurant = require('./schema_models/restaurantSchema.js');
 
 const generateUsers = async (count = 50) => {
+    console.log('Generating new users')
     try {
         await User.deleteMany({});
         console.log('Old users cleared.');
@@ -22,7 +23,7 @@ const generateUsers = async (count = 50) => {
                 password: hashedPass,
                 avatar: faker.image.avatar(),
                 description: faker.person.bio(),
-                role: faker.helpers.arrayElement(['user', 'owner']),
+                role: 'user',
             });
         }
 
@@ -34,15 +35,56 @@ const generateUsers = async (count = 50) => {
     }
 };
 
-const generateRestaurants = async (count = 50) => {
+const generateOwners = async (count = 10) => {
+    console.log('Generating new owner users (note: this must be run AFTER generating users otherwise you will have a bad time)')
+    try {
+        let password = 'password'
+        let saltRounds = 10;
+
+        const users = [];
+        for (let i = 0; i < count; i++) {
+            let hashedPass = bcrpyt.hashSync(password, saltRounds)
+            users.push({
+                username: faker.internet.username(),
+                password: hashedPass,
+                avatar: faker.image.avatar(),
+                description: faker.person.bio(),
+                role: 'owner',
+            });
+        }
+
+        await User.insertMany(users);
+        console.log(`Successfully created ${count} owner users!`);
+        return users;
+
+    } catch (error) {
+        console.error('Error generating users: ', error);
+    }
+};
+
+const generateRestaurants = async (ownerList, count = 10) => {
+    if (ownerList.length < count) {
+        console.log(`Provided owner list is too short ${ownerList.length} to generate ${count} restaurants`)
+        return
+    }
+
     try {
         await Restaurant.deleteMany({});
         console.log('Old restaurants cleared');
+
         
         const restaurants = [];
         for (let i=0; i<count; i++) {
             const minPrice = parseFloat(faker.commerce.price({ min: 80, max: 400 }));
             const maxPrice = parseFloat(faker.commerce.price({ min: Math.floor(minPrice * 1.5), max: 900 }));
+
+            const qry = User.find({username: ownerList[i].username })
+                .select('_id')
+                .lean();
+            
+            const ownerId = await qry.exec();
+            // console.log(ownerId[0]._id);
+
             restaurants.push({
                 name: `${faker.company.name()} ${faker.helpers.arrayElement(['Kitchen', 'Bistro', 'Grill', 'Cafe', 'Diner'])}`,
                 description: faker.lorem.sentences(2),
@@ -52,10 +94,11 @@ const generateRestaurants = async (count = 50) => {
                     province: faker.location.state(),
                     zipCode: faker.location.zipCode()
                 },
+                ownerId: ownerId[0]._id,
                 priceRange: {
                     min: minPrice,
                     max: maxPrice
-                }
+                },
             });
         }
 
@@ -69,9 +112,13 @@ const generateRestaurants = async (count = 50) => {
 
 const generateData = async () => {
     console.log('Connecting to database...')
+    const userCount = 50;
+    const rstCount = 10;
+
     await connectDB()
-    await generateUsers(50) 
-    await generateRestaurants(50)
+    await generateUsers(userCount) 
+    const owners = await generateOwners(rstCount)
+    await generateRestaurants(owners, rstCount)
 
     process.exit(0)
 }
